Add interfaces for access and refresh token payloads; restructure auth and session services
