+++ backend/src/auth/auth.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './services/auth.service';
 import { AuthController } from './controllers/auth.controller';
 import { UserModule } from 'src/user/user.module';
-import { BcryptService } from './services/bcrypt.service';
-import { CryptoService } from './services/crypto.service';
+import { BcryptService } from './providers/bcrypt.service';
+import { CryptoService } from './providers/crypto.service';
 import { SessionModule } from 'src/session/session.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
