+++ backend/src/auth/index.d.ts
@@ -1,0 +1,10 @@
+export interface AccessTokenPayload{
+    userId: string;
+    deviceId: string;
+}
+
+export interface RefreshTokenPayload{
+    userId: string;
+    deviceId: string;
+    refreshToken: string;
+}
